quiz/requirements.txt
aioredis==1.3.1 
APScheduler==3.10.1 
asgiref==3.4.1 
async-timeout==4.0.2 
attrs==22.2.0 
autobahn==21.2.1 
Automat==22.10.0 
backports.zoneinfo==0.2.1 
certifi==2022.12.7 
cffi==1.15.1 
channels==3.0.4 
channels-redis==3.3.1 
charset-normalizer==2.0.12 
click==8.0.4 
colorama==0.4.5 
constantly==15.1.0 
cryptography==39.0.2 
daphne==3.0.2 
dataclasses==0.8 
Django==3.2.16 
django-admin-interface==0.23.0 
django-colorfield==0.7.3 
django-cors-headers==3.10.1 
django-crontab==0.7.1 
django-extra-fields==3.0.2 
django-filter==21.1 
django-flat-responsive==2.0 
django-flat-theme==1.1.4 
django-mysql==4.3.0 
django-rest-knox==4.2.0 
django-sslserver==0.22 
djangorestframework==3.14.0 
h11==0.13.0 
hiredis==2.0.0 
httpagentparser==1.9.5 
httptools==0.3.0 
hyperlink==21.0.0 
idna==3.4 
importlib-metadata==4.8.3 
importlib-resources==5.4.0 
incremental==22.10.0 
Jinja2==3.0.3 
MarkupSafe==2.0.1 
msgpack==1.0.5 
mysql-connector-python==8.0.31 
packaging==21.3 
Pillow==8.4.0 
protobuf==3.19.6 
pyasn1==0.4.8 
pyasn1-modules==0.2.8 
pycparser==2.21 
PyMySQL==1.4.3 
pyOpenSSL==23.0.0 
pyparsing==3.0.9 
python-dotenv==0.20.0 
pytz==2022.7.1 
pytz-deprecation-shim==0.1.0.post0 
PyYAML==6.0 
redis==4.3.5 
requests==2.27.1 
service-identity==21.1.0 
six==1.16.0 
sqlparse==0.4.3 
Twisted==22.4.0 
twisted-iocpsupport==1.0.2 
txaio==22.2.1 
typing_extensions==4.1.1 
tzdata==2022.7 
tzlocal==4.2 
urllib3==1.26.15 
uvicorn==0.16.0 
watchgod==0.7 
websockets==9.1 
zipp==3.6.0 
zope.interface==5.5.2